Removes Onyxbay. Adds Aurora

+++ code/datums/trading/trade.dm
@@ -12,7 +12,8 @@
 	var/list/possible_trading_items                             //List of all possible trading items. Structure is (type = mode)
 	var/list/trading_items = list()                             //What items they are currently trading away.
 	var/list/blacklisted_trade_items = list(/mob/living/carbon/human)
-	                                                            //Things they will automatically refuse
+	var/list/allowed_space_sectors = list(SECTOR_ROMANOVICH, SECTOR_TAU_CETI, SECTOR_CORP_ZONE, SECTOR_VALLEY_HALE, SECTOR_BADLANDS, SECTOR_NEW_ANKARA, SECTOR_AEMAQ, SECTOR_SRANDMARR, SECTOR_NRRAHRAHUL,
+										SECTOR_GAKAL, SECTOR_UUEOAESA)	//which sector this merchant can show up                                                            //Things they will automatically refuse
 
 	var/list/speech = list()                                    //The list of all their replies and messages. Structure is (id = talk)
 	/*SPEECH IDS:
@@ -33,7 +34,6 @@
 
 	*/
 	var/want_multiplier = 2                                     //How much wanted items are multiplied by when traded for
-	var/margin = 1  											//Multiplier to price when selling to player
 	var/insult_drop = 5                                         //How far disposition drops on insult
 	var/compliment_increase = 5                                 //How far compliments increase disposition
 	var/refuse_comms = 0                                        //Whether they refuse further communication
@@ -44,7 +44,7 @@
 	..()
 	if(name_language)
 		if(name_language == TRADER_DEFAULT_NAME)
-			name = capitalize(pick(GLOB.first_names_female + GLOB.first_names_male)) + " " + capitalize(pick(GLOB.last_names))
+			name = capitalize(pick(first_names_female + first_names_male)) + " " + capitalize(pick(last_names))
 		else
 			var/datum/language/L = all_languages[name_language]
 			if(L)
@@ -52,48 +52,28 @@
 	if(possible_origins && possible_origins.len)
 		origin = pick(possible_origins)
 
-	if(possible_wanted_items)
-		possible_wanted_items = generate_pool(possible_wanted_items)
-	if(possible_trading_items)
-		possible_trading_items = generate_pool(possible_trading_items)
-
 	for(var/i in 5 to 8)
 		add_to_pool(trading_items, possible_trading_items, force = 1)
 		add_to_pool(wanted_items, possible_wanted_items, force = 1)
 
-/datum/trader/proc/generate_pool(list/trading_pool)
-	. = list()
-	// Add types
-	for(var/type in trading_pool)
-		var/status = trading_pool[type]
-		if(status & TRADER_THIS_TYPE)
-			. += type
-		if(status & TRADER_SUBTYPES_ONLY)
-			. += subtypesof(type)
-
-	// Remove blacklisted
-	for(var/type in .)
-		var/status = trading_pool[type]
-		if(status & TRADER_BLACKLIST)
-			. -= type
-		if(status & TRADER_BLACKLIST_SUB)
-			. -= subtypesof(type)
-
 //If this hits 0 then they decide to up and leave.
 /datum/trader/proc/tick()
-	spawn(0)
-		add_to_pool(trading_items, possible_trading_items, 200)
-		add_to_pool(wanted_items, possible_wanted_items, 50)
-		remove_from_pool(possible_trading_items, 9) //We want the stock to change every so often, so we make it so that they have roughly 10~11 ish items max
+	addtimer(CALLBACK(src, PROC_REF(do_after_tick)), 1)
 	return 1
 
-/datum/trader/proc/remove_from_pool(list/pool, chance_per_item)
+/datum/trader/proc/do_after_tick()
+	add_to_pool(trading_items, possible_trading_items, 200)
+	add_to_pool(wanted_items, possible_wanted_items, 50)
+	remove_from_pool(possible_trading_items, 9) //We want the stock to change every so often, so we make it so that they have roughly 10~11 ish items max
+	return
+
+/datum/trader/proc/remove_from_pool(var/list/pool, var/chance_per_item)
 	if(pool && prob(chance_per_item * pool.len))
 		var/i = rand(1,pool.len)
 		pool[pool[i]] = null
 		pool -= pool[i]
 
-/datum/trader/proc/add_to_pool(list/pool, list/possible, base_chance = 100, force = 0)
+/datum/trader/proc/add_to_pool(var/list/pool, var/list/possible, var/base_chance = 100, var/force = 0)
 	var/divisor = 1
 	if(pool && pool.len)
 		divisor = pool.len
@@ -102,17 +82,29 @@
 		if(new_item)
 			pool |= new_item
 
-/datum/trader/proc/get_possible_item(list/trading_pool)
+/datum/trader/proc/get_possible_item(var/list/trading_pool)
 	if(!trading_pool || !trading_pool.len)
 		return
-	var/picked = pick(trading_pool)
-	var/atom/A = picked
-	// Weed out a few of the common bad types. Reason we don't check types specifically is that (hopefully) further bad subtypes don't set their name up and are similar.
-	if(initial(A.name) in list("object", "item","weapon", "structure", "machinery", "Mecha", "organ", "snack"))
-		return null
-	return picked
+	var/list/possible = list()
+	for(var/type in trading_pool)
+		var/status = trading_pool[type]
+		if(status & TRADER_THIS_TYPE)
+			possible += type
+		if(status & TRADER_SUBTYPES_ONLY)
+			possible += subtypesof(type)
+		if(status & TRADER_BLACKLIST)
+			possible -= type
+		if(status & TRADER_BLACKLIST_SUB)
+			possible -= subtypesof(type)
 
-/datum/trader/proc/get_response(key, default)
+	if(possible.len)
+		var/picked = pick(possible)
+		var/atom/A = picked
+		if(initial(A.name) in list("object", "item","weapon", "structure", "machinery", "exosuit", "organ", "snack")) //weed out a few of the common bad types. Reason we don't check types specifically is that (hopefully) further bad subtypes don't set their name up and are similar.
+			return
+		return picked
+
+/datum/trader/proc/get_response(var/key, var/default)
 	var/text
 	if(speech && speech[key])
 		text = speech[key]
@@ -121,71 +113,68 @@
 	text = replacetext(text, "MERCHANT", name)
 	return replacetext(text, "ORIGIN", origin)
 
-/datum/trader/proc/print_trading_items(num)
+/datum/trader/proc/print_trading_items(var/num)
 	num = Clamp(num,1,trading_items.len)
 	if(trading_items[num])
 		var/atom/movable/M = trading_items[num]
-		return "<b>[initial(M.name)]</b>"
+		return "[initial(M.name)]"
 
-/datum/trader/proc/get_item_value(trading_num)
+/datum/trader/proc/get_item_value(var/trading_num)
 	if(!trading_items[trading_items[trading_num]])
 		var/type = trading_items[trading_num]
 		var/value = get_value(type)
 		value = round(rand(80,100)/100 * value) //For some reason rand doesn't like decimals.
-		trading_items[type] = margin*value
+		trading_items[type] = value
 	return trading_items[trading_items[trading_num]]
 
-/datum/trader/proc/make_response(response_type, response_default, delta = 0, success = TRUE)
-	. = new /datum/trade_response(get_response(response_type, response_default), delta, success)
-
-/datum/trader/proc/offer_money_for_trade(trade_num, money_amount, location)
+/datum/trader/proc/offer_money_for_trade(var/trade_num, var/money_amount)
 	if(!(trade_flags & TRADER_MONEY))
-		return make_response(TRADER_NO_MONEY, "I don't like money", 0, FALSE)
+		return TRADER_NO_MONEY
 	var/value = get_item_value(trade_num)
 	if(money_amount < value)
-		return make_response(TRADER_NOT_ENOUGH, "I need more money!", 0, FALSE)
-	trade(list(), trade_num, location)
-	return make_response(TRADER_TRADE_COMPLETE, "Thank you for your patronage!", -value, TRUE)
+		return TRADER_NOT_ENOUGH
 
-/datum/trader/proc/offer_items_for_trade(list/offers, num, turf/location)
+	return value
+
+/datum/trader/proc/offer_items_for_trade(var/list/offers, var/num, var/turf/location)
 	if(!offers || !offers.len)
-		return make_response(TRADER_NOT_ENOUGH, "That's not enough.", 0, FALSE)
+		return TRADER_NOT_ENOUGH
 	num = Clamp(num, 1, trading_items.len)
 	var/offer_worth = 0
 	for(var/item in offers)
 		var/atom/movable/offer = item
 		var/is_wanted = 0
-		if((trade_flags & TRADER_WANTED_ONLY) && is_type_in_list(offer,wanted_items))
-			is_wanted = 2
-		if((trade_flags & TRADER_WANTED_ALL) && is_type_in_list(offer,possible_wanted_items))
+		if(is_type_in_list(offer,wanted_items))
 			is_wanted = 1
-		if(blacklisted_trade_items && blacklisted_trade_items.len && is_type_in_list(offer,blacklisted_trade_items))
-			return make_response(TRADER_NO_BLACKLISTED, "I refuse to take one of those items.", 0, FALSE)
+		if(blacklisted_trade_items && blacklisted_trade_items.len)
+			if(ishuman(offer))
+				var/mob/living/carbon/human/A = offer
+				if(is_type_in_list(A.species, blacklisted_trade_items))
+					return 0
+			else if(is_type_in_list(offer,blacklisted_trade_items))
+				return 0
 
 		if(istype(offer,/obj/item/spacecash))
 			if(!(trade_flags & TRADER_MONEY))
-				return make_response(TRADER_NO_MONEY, "I don't take money.", 0, FALSE)
+				return TRADER_NO_MONEY
 		else
 			if(!(trade_flags & TRADER_GOODS))
-				return make_response(TRADER_NO_GOODS, "I don't take goods.", 0, FALSE)
-			else if((trade_flags & TRADER_WANTED_ONLY|TRADER_WANTED_ALL) && !is_wanted)
-				return make_response(TRADER_FOUND_UNWANTED, "I don't want one of those items", 0, FALSE)
+				return TRADER_NO_GOODS
+			else if((trade_flags & TRADER_WANTED_ONLY) && !is_wanted)
+				return TRADER_FOUND_UNWANTED
 
-		offer_worth += get_value(offer) * (is_wanted > 1 ? want_multiplier : 1)
+		offer_worth += get_value(offer) * (is_wanted ? want_multiplier : 1)
 	if(!offer_worth)
-		return make_response(TRADER_NOT_ENOUGH, "Everything you've gave me is worthless!", 0, FALSE)
+		return TRADER_NOT_ENOUGH
 	var/trading_worth = get_item_value(num)
 	if(!trading_worth)
-		return make_response(TRADER_NOT_ENOUGH, "That's not enough.", 0, FALSE)
+		return TRADER_NOT_ENOUGH
 	var/percent = offer_worth/trading_worth
 	if(percent > max(0.9,0.9-disposition/100))
-		trade(offers, num, location)
-		return make_response(TRADER_TRADE_COMPLETE, "Thank you for your patronage!", 0, TRUE)
-	return make_response(TRADER_NOT_ENOUGH, "That's not enough.", 0, FALSE)
+		return trade(offers, num, location)
+	return TRADER_NOT_ENOUGH
 
-/datum/trader/proc/hail(mob/user)
-	if(!can_hail())
-		return make_response(TRADER_HAIL_DENY, "No, go away.", 0, FALSE)
+/datum/trader/proc/hail(var/mob/user)
 	var/specific
 	if(istype(user, /mob/living/carbon/human))
 		var/mob/living/carbon/human/H = user
@@ -193,11 +182,10 @@
 			specific = H.species.name
 	else if(istype(user, /mob/living/silicon))
 		specific = "silicon"
-	if(!speech[TRADER_HAIL_START + specific])
+	if(!speech["hail_[specific]"])
 		specific = "generic"
-	var/datum/trade_response/tr = make_response(TRADER_HAIL_START + specific, "Greetings, MOB!", 0, TRUE)
-	tr.text = replacetext(tr.text, "MOB", user.name)
-	return tr
+	. = get_response("hail_[specific]", "Greetings, MOB!")
+	. = replacetext(., "MOB", user.name)
 
 /datum/trader/proc/can_hail()
 	if(!refuse_comms && prob(-disposition))
@@ -209,81 +197,72 @@
 	if(prob(-disposition/10))
 		refuse_comms = 1
 	if(disposition > 50)
-		return make_response(TRADER_INSULT_GOOD,"What? I thought we were cool!", 0, TRUE)
+		return get_response("insult_good","What? I thought we were cool!")
 	else
-		return make_response(TRADER_INSULT_BAD, "Right back at you asshole!", 0, FALSE)
+		return get_response("insult_bad", "Right back at you asshole!")
 
 /datum/trader/proc/compliment()
 	if(prob(-disposition))
-		return make_response(TRADER_COMPLEMENT_FAILURE, "Fuck you!", 0, FALSE)
+		return get_response("compliment_deny", "Fuck you!")
 	if(prob(100-disposition))
 		disposition += rand(compliment_increase, compliment_increase * 2)
-	return make_response(TRADER_COMPLEMENT_SUCCESS, "Thank you!", 0, TRUE)
+	return get_response("compliment_accept", "Thank you!")
 
-/datum/trader/proc/trade(list/offers, num, turf/location)
-	for(var/offer in offers)
-		if(istype(offer, /mob))
-			var/mob/M = offer
-			var/text = mob_transfer_message
-			to_chat(M, replacetext(text, "ORIGIN", origin))
-			if(M.key)
-				M.ghostize()
-		if(istype(offer, /obj/mecha))
-			var/obj/mecha/M = offer
-			M.wreckage = null //So they don't ruin the illusion
-		qdel(offer)
+/datum/trader/proc/trade(var/list/offers, var/num, var/turf/location)
+	if(offers && offers.len)
+		for(var/offer in offers)
+			if(istype(offer,/mob))
+				var/text = mob_transfer_message
+				to_chat(offer, replacetext(text, "ORIGIN", origin))
+			qdel(offer)
 
-	num = Clamp(num, 1, trading_items.len)
 	var/type = trading_items[num]
 
 	var/atom/movable/M = new type(location)
-	M.on_purchase()
 	playsound(location, 'sound/effects/teleport.ogg', 50, 1)
 
 	disposition += rand(compliment_increase,compliment_increase*3) //Traders like it when you trade with them
 
 	return M
 
-/datum/trader/proc/how_much_do_you_want(num)
-	num = Clamp(num, 1, trading_items.len)
+/datum/trader/proc/how_much_do_you_want(var/num)
 	var/atom/movable/M = trading_items[num]
-	var/datum/trade_response/tr = make_response(TRADER_HOW_MUCH, "Hmm.... how about VALUE CURRENCY?", 0, FALSE)
-	tr.text = replacetext(replacetext(tr.text, "ITEM", initial(M.name)), "VALUE", get_item_value(num))
-	return tr
+	. = get_response("how_much", "Hmm.... how about VALUE credits?")
+	. = replacetext(.,"VALUE",get_item_value(num))
+	. = replacetext(.,"ITEM", initial(M.name))
 
 /datum/trader/proc/what_do_you_want()
 	if(!(trade_flags & TRADER_GOODS))
-		return make_response(TRADER_NO_GOODS, "I don't deal in goods.", 0, FALSE)
+		return get_response(TRADER_NO_GOODS, "I don't deal in goods.")
 
-	var/datum/trade_response/tr = make_response(TRADER_WHAT_WANT, "Hm, I want", 0, TRUE)
+	. = get_response("what_want", "Hm, I want")
 	var/list/want_english = list()
 	for(var/type in wanted_items)
 		var/atom/a = type
 		want_english += initial(a.name)
-	tr.text += " [english_list(want_english)]"
-	return tr
+	. += " [english_list(want_english)]"
 
-/datum/trader/proc/sell_items(list/offers)
+/datum/trader/proc/sell_items(var/list/offers)
 	if(!(trade_flags & TRADER_GOODS))
-		return make_response(TRADER_GOODS, "I'm not buying.", 0, FALSE)
+		return TRADER_NO_GOODS
 	if(!offers || !offers.len)
-		return make_response(TRADER_NOT_ENOUGH, "I'm not buying that.", 0, FALSE)
+		return TRADER_NOT_ENOUGH
 
-	var/mult = 1
-	var/total = 0
 	for(var/offer in offers)
-		if((trade_flags & TRADER_WANTED_ONLY) && is_type_in_list(offer,wanted_items))
-			mult = want_multiplier
-		else if((trade_flags & TRADER_WANTED_ALL) && is_type_in_list(offer,possible_wanted_items))
-			mult = 1
-		else
-			return make_response(TRADER_FOUND_UNWANTED, "I don't want one of those items", 0, FALSE)
-		total += get_value(offer) * mult
+		if(!is_type_in_list(offer,wanted_items))
+			return TRADER_FOUND_UNWANTED
 
-	playsound(offers[1], 'sound/effects/teleport.ogg', 50, 1)
+	playsound(get_turf(offers[1]), 'sound/effects/teleport.ogg', 50, 1)
+	. = 0
 	for(var/offer in offers)
+		. += get_value(offer) * want_multiplier
 		qdel(offer)
-	return make_response(TRADER_TRADE_COMPLETE, "Thanks for the goods!", total, TRUE)
 
-/datum/trader/proc/bribe_to_stay_longer(amt)
-	return make_response(TRADER_BRIBE_FAILURE, "How about no?", 0, FALSE)
+/datum/trader/proc/bribe_to_stay_longer(var/amt)
+	return get_response("bribe_refusal", "How about... no?")
+
+/datum/trader/proc/system_allowed()
+	if(SSatlas.current_sector.name in allowed_space_sectors)
+		return TRUE
+	else
+		return FALSE
